backend/requirements-dev.txt
# Development dependencies
black>=23.0
ruff>=0.1.3
pytest>=7.0
pytest-django>=4.5
ipython>=8.0  # for better shell experience
watchdog>=3.0  # for auto-reload